requirements.txt
pvporcupine==2.2.1
PyAudio==0.2.13
python-dotenv==1.0.0

openai~=0.27.10
numpy~=1.24.4
scipy~=1.11.2
gTTS~=2.3.2
librosa~=0.10.1
pydub~=0.25.1